refactor: update COP conference version in dialog descriptions
Update the conference version in the dialog descriptions for both English and Portuguese translations. Change "COP19" to "COP29" to reflect the correct conference edition.

Also, update the Sonner component theme and add the richColors prop, as well as update the AboutBasa component and VideoPlayer.

Reset the form after successful redemption in the Hero component.

Fix a typo in the Pertners component, changing "COP19" to "COP29" to match the updated conference version.